app_streamlit/requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
pandas==2.2.2
plotly==5.23.0
